fix: address review of the scaler and GL path
Three fixes from the PR review:

The exact-indexing rewrite carried `src` and subtracted `dst` until it
fit, which is `src / dst` iterations per pixel -- 220M for the 2.2M -> 100
case the tests cover, against 10k before. Precompute the quotient and
remainder so each pixel costs one add and at most one carry. Same output,
measured 39ms -> 0.01ms for that blit.

Require highp in the fragment language rather than falling back to
mediump. mediump guarantees 2^-10 relative precision, which cannot
resolve adjacent texels above ~1024 wide, so the fallback sampled the
wrong texel on buffers GL_MAX_TEXTURE_SIZE accepts. #error makes the
shader ill-formed, so such a driver takes the software path instead.

Take the buffer dimensions as usize in present() and convert with
try_from. Casting wrapped a width past i32::MAX to a negative one, which
slipped under the size check, skipped the upload and reported success --
presenting a cleared window instead of falling back. A zero height lets
any width through check_buffer_size, so it was reachable from safe code.

+++ src/os/posix/wayland.rs
@@ -802,7 +802,7 @@ impl Window {
             // Still too big for the context we would rebuild; do not pay for
             // one just to hit the same limit.
             GlPath::TooLarge { max } => {
-                if buf_width as i32 > max || buf_height as i32 > max {
+                if buf_width > max as usize || buf_height > max as usize {
                     return false;
                 }
             }
@@ -839,9 +839,9 @@ impl Window {
         let result = unsafe {
             context.context().present(
                 buffer,
-                buf_width as i32,
-                buf_height as i32,
-                buf_stride as i32,
+                buf_width,
+                buf_height,
+                buf_stride,
                 size.width,
                 size.height,
                 scale_mode,
